requirements.txt
langchain
langchain-groq
python-dotenv
playwright
opencv-python
numpy